Introduction to Econometrics
This Github repository will give an overview and introduction to econometrics. In the first part of the repository, preliminaries such as matrix algebra and statistics will be briefly covered, as well as classical linear regression and basic regression tests. The second part of the repository will focus on the application of these concepts. It will delve into regression diagnostic tests and time series analysis.